added Initial Windows Mobile 6 Version. - Windows Mobile 6 SDK - Visual Studio 2005 only Minor: - Burningvideo: MipMap Selection repaired - renamed private Driver function getTextureSizeFromImageSize to   getTextureSizeFromSurfaceSize
git-svn-id: https://irrlicht.svn.sourceforge.net/svnroot/irrlicht/trunk@1146 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/quaternion.h
@@ -483,7 +483,7 @@ inline void quaternion::toAngleAxis(f32 &angle, core::vector3df &axis) const
 	}
 	else
 	{
-		angle = 2.0f * acos(W);
+		angle = 2.0f * (f32) acos(W);
 		axis.X = X / scale;
 		axis.Y = Y / scale;
 		axis.Z = Z / scale;
